0.8
add edite profile
fix hike users in show hike
fix table users and hike users

+++ database/migrations/2006_01_12_000000_create_users_table.php
@@ -25,7 +25,7 @@ class CreateUsersTable extends Migration
             $table->string('gender');
             $table->text('img')->nullable();
             $table->date('birth_date')->nullable();
-            $table->integer('telephone')->nullable();
+            $table->bigInteger('telephone')->nullable();
             $table->text('info')->nullable();
             $table->rememberToken();
             $table->timestamps();
